NewtonBlackølInterationInterleaved: NewtonBlackølInterationInterleavedImpl implements the Iteration for a fixed number of cell variabled which is then used by the NewtonBlackølInterationInterleaved class via a switch-case over the actually existing numbers.
